@@ -3,19 +3,18 @@ package danielvishnievskyi.bachelorproject.entities;
 import com.fasterxml.jackson.annotation.JsonIdentityInfo;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.voodoodyne.jackson.jsog.JSOGGenerator;
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 import org.hibernate.envers.AuditMappedBy;
 import org.hibernate.envers.Audited;
-import org.springframework.data.envers.repository.config.EnableEnversRepositories;
 
-import javax.persistence.*;
 import java.util.Collection;
 
-import static javax.persistence.CascadeType.MERGE;
-import static javax.persistence.FetchType.EAGER;
+import static jakarta.persistence.CascadeType.MERGE;
+import static jakarta.persistence.FetchType.EAGER;
 
 @Entity
 @Setter
