@@ -6,6 +6,8 @@ import danielvishnievskyi.bachelorproject.entities.AdminProfile;
 import danielvishnievskyi.bachelorproject.entities.Role;
 import danielvishnievskyi.bachelorproject.security.utils.AuthUtil;
 import danielvishnievskyi.bachelorproject.services.AdminProfileService;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
@@ -14,8 +16,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
