@@ -1,9 +1,9 @@
 package danielvishnievskyi.bachelorproject.dto;
 
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
 
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotBlank;
 
 @Data
 public class BuildingDto {
