@@ -1,8 +1,8 @@
 package danielvishnievskyi.bachelorproject.dto;
 
+import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
 
-import javax.validation.constraints.NotBlank;
 
 @Data
 public class LocationDto {
