@@ -2,13 +2,13 @@ package danielvishnievskyi.bachelorproject.entities;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import danielvishnievskyi.bachelorproject.implementations.EnversRevImpl;
+import jakarta.persistence.Entity;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 import org.hibernate.envers.DefaultRevisionEntity;
 import org.hibernate.envers.RevisionEntity;
 
-import javax.persistence.Entity;
 
 @Entity
 @Getter
