@@ -3,10 +3,10 @@ package danielvishnievskyi.bachelorproject.implementations;
 import danielvishnievskyi.bachelorproject.entities.Privilege;
 import danielvishnievskyi.bachelorproject.entities.Role;
 import danielvishnievskyi.bachelorproject.repositories.RoleRepo;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
-import javax.transaction.Transactional;
 import java.util.Collection;
 
 @Component
