@@ -2,12 +2,12 @@ package danielvishnievskyi.bachelorproject.repositories.specifications;
 
 import danielvishnievskyi.bachelorproject.entities.Building;
 import danielvishnievskyi.bachelorproject.repositories.criteria.SearchCriteria;
+import jakarta.persistence.criteria.CriteriaBuilder;
+import jakarta.persistence.criteria.CriteriaQuery;
+import jakarta.persistence.criteria.Predicate;
+import jakarta.persistence.criteria.Root;
 import org.springframework.data.jpa.domain.Specification;
 
-import javax.persistence.criteria.CriteriaBuilder;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Predicate;
-import javax.persistence.criteria.Root;
 import java.util.ArrayList;
 import java.util.List;
 
